Renamed Ddl.Parser.Tests test project folder
