+++ tests/Ddl.Parser.Tests/AssertExceptions/AssertLexerException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace TheToolsmiths.Ddl.Parser.Tests.AssertExceptions
+{
+    public class AssertLexerException : Exception
+    {
+        public AssertLexerException(string message)
+            : base(message)
+        {
+        }
+    }
+}
