+++ tests/Ddl.Parser.Tests/AssertExceptions/AssertParserException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace TheToolsmiths.Ddl.Parser.Tests.AssertExceptions
+{
+    public class AssertParserException : Exception
+    {
+        public AssertParserException(string message)
+            : base(message)
+        {
+        }
+    }
+}
