+++ tests/Ddl.Parser.Tests/Utils/PathConstants.cs
@@ -1,0 +1,7 @@
+﻿namespace TheToolsmiths.Ddl.Parser.Tests.Utils
+{
+    public static class PathConstants
+    {
+        public static string ExamplesFolder { get; } = "Examples";
+    }
+}
