--- tests/DdlParser.Tests/AssertExceptions/AssertLexerException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace TheToolsmiths.Ddl.Parser.Tests.AssertExceptions
-{
-    public class AssertLexerException : Exception
-    {
-        public AssertLexerException(string message)
-            : base(message)
-        {
-        }
-    }
-}
