--- tests/DdlParser.Tests/AssertExceptions/AssertParserException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace TheToolsmiths.Ddl.Parser.Tests.AssertExceptions
-{
-    public class AssertParserException : Exception
-    {
-        public AssertParserException(string message)
-            : base(message)
-        {
-        }
-    }
-}
