--- tests/DdlParser.Tests/Ddl.Parser.Tests.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29411.108
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ddl.Parser.Tests", "Ddl.Parser.Tests.csproj", "{A98C3DD7-37F8-4D03-86E0-DEEDF83ED8F3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ddl.Parser", "..\..\src\Ddl.Parser\Ddl.Parser.csproj", "{7F72BB6A-AB0C-4B96-B7C8-E9690B00F66A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A98C3DD7-37F8-4D03-86E0-DEEDF83ED8F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A98C3DD7-37F8-4D03-86E0-DEEDF83ED8F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A98C3DD7-37F8-4D03-86E0-DEEDF83ED8F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A98C3DD7-37F8-4D03-86E0-DEEDF83ED8F3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7F72BB6A-AB0C-4B96-B7C8-E9690B00F66A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7F72BB6A-AB0C-4B96-B7C8-E9690B00F66A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7F72BB6A-AB0C-4B96-B7C8-E9690B00F66A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7F72BB6A-AB0C-4B96-B7C8-E9690B00F66A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C7426B21-71D8-494D-BB84-2E5E25E46321}
-	EndGlobalSection
-EndGlobal
