--- tests/DdlParser.Tests/Ddl.Parser.Tests.sln.DotSettings
@@ -1,10 +1,0 @@
-﻿<wpf:ResourceDictionary xml:space="preserve" 
-	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
-	xmlns:s="clr-namespace:System;assembly=mscorlib" 
-	xmlns:ss="urn:shemas-jetbrains-com:settings-storage-xaml" 
-	xmlns:wpf="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
-	<s:String x:Key="/Default/Environment/InjectedLayers/FileInjectedLayer/=6D144E251817D84584A4AD048EAD2F7B/RelativePath/@EntryValue">..\..\..\src\Ddl.Parser.sln.DotSettings</s:String>
-	<s:Boolean x:Key="/Default/Environment/InjectedLayers/FileInjectedLayer/=6D144E251817D84584A4AD048EAD2F7B/@KeyIndexDefined">True</s:Boolean>
-	<s:Boolean x:Key="/Default/Environment/InjectedLayers/InjectedLayerCustomization/=File6D144E251817D84584A4AD048EAD2F7B/@KeyIndexDefined">True</s:Boolean>
-	<s:Double x:Key="/Default/Environment/InjectedLayers/InjectedLayerCustomization/=File6D144E251817D84584A4AD048EAD2F7B/RelativePriority/@EntryValue">1</s:Double>
-</wpf:ResourceDictionary>
