--- tests/DdlParser.Tests/Utils/PathConstants.cs
@@ -1,7 +1,0 @@
-﻿namespace TheToolsmiths.Ddl.Parser.Tests.Utils
-{
-    public static class PathConstants
-    {
-        public static string ExamplesFolder { get; } = "Examples";
-    }
-}
